requirements.txt
streamlit==1.22.0
pandas==1.5.3
streamlit-card==0.0.4
requests==2.28.1
beautifulsoup4==4.11.1
nltk==3.7
tensorflow-cpu==2.11.0
numpy==1.23.5
keras==2.11.0
